// src/models/consent.model.ts
import mongoose, { Schema, Document } from "mongoose";

export interface IConsent extends Document {
  id: string;
  name: string;
  consentUrl: string;
  createdAt: Date;
  version: number;
}

const ConsentSchema: Schema = new Schema(
  {
    // The id of the target, generated by api in UUID
    id: { type: String },

    // A human readable name of the consent we are granting to the user. Required field.
    name: { type: String, required: true },

    // The url. Required field
    consentUrl: { type: String, required: true },
    version: { type: Number },
  },
  { timestamps: true }
);

export default mongoose.model<IConsent>("Consent", ConsentSchema);